+++ src/options/presets/Default.ts
@@ -44,7 +44,7 @@ export const DEFAULT_PRESET: TInputOptions = Object.freeze({
     stringArrayEncoding: [
         StringArrayEncoding.None
     ],
-    stringArrayIntermediateCalls: 0,
+    stringArrayIntermediateVariablesCount: 0,
     stringArrayThreshold: 0.75,
     target: ObfuscationTarget.Browser,
     transformObjectKeys: false,