+++ test/functional-tests/options/OptionsNormalizer.spec.ts
@@ -544,7 +544,7 @@ describe('OptionsNormalizer', () => {
                     shuffleStringArray: true,
                     stringArray: false,
                     stringArrayEncoding: [StringArrayEncoding.Rc4],
-                    stringArrayIntermediateCalls: 5,
+                    stringArrayIntermediateVariablesCount: 5,
                     stringArrayThreshold: 0.5,
                     rotateStringArray: true
                 });
@@ -554,7 +554,7 @@ describe('OptionsNormalizer', () => {
                     shuffleStringArray: false,
                     stringArray: false,
                     stringArrayEncoding: [StringArrayEncoding.None],
-                    stringArrayIntermediateCalls: 0,
+                    stringArrayIntermediateVariablesCount: 0,
                     stringArrayThreshold: 0,
                     rotateStringArray: false
                 };
@@ -592,7 +592,7 @@ describe('OptionsNormalizer', () => {
                     rotateStringArray: true,
                     shuffleStringArray: true,
                     stringArray: true,
-                    stringArrayIntermediateCalls: 5,
+                    stringArrayIntermediateVariablesCount: 5,
                     stringArrayThreshold: 0
                 });
 
@@ -601,7 +601,7 @@ describe('OptionsNormalizer', () => {
                     rotateStringArray: false,
                     shuffleStringArray: false,
                     stringArray: false,
-                    stringArrayIntermediateCalls: 0,
+                    stringArrayIntermediateVariablesCount: 0,
                     stringArrayThreshold: 0
                 };
             });