+++ test/runtime-tests/JavaScriptObfuscatorRuntime.spec.ts
@@ -39,7 +39,7 @@ describe('JavaScriptObfuscator runtime eval', function () {
             StringArrayEncoding.Base64,
             StringArrayEncoding.Rc4
         ],
-        stringArrayIntermediateCalls: 20,
+        stringArrayIntermediateVariablesCount: 20,
         stringArrayThreshold: 1,
         transformObjectKeys: true,
         unicodeEscapeSequence: true